requirements.txt
duckdb==1.1.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.31.0
pandas==2.2.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
tqdm==4.66.1
